Made the hash generation as separate functionality due to circular import issue
