+++ common/api/core-backend.api.md
@@ -3592,8 +3592,10 @@ export class Schema {
     static get schemaName(): string;
     // @beta
     static toSemverString(paddedVersion: string): string;
-    static getSchemaSha1Hash(schemaXmlPath: string, referencePaths: string[], isExactMatch: boolean = false): string;
 }
+
+// @public
+export function getSchemaSha1Hash(schemaXmlPath: string, referencePaths: string[], isExactMatch: boolean = false): string;
 
 // @internal (undocumented)
 export type SchemaKey = IModelJsNative.ECSchemaXmlContext.SchemaKey;
